requirements.txt
selenium==4.17.2
pytest==8.0.0
allure-pytest==2.13.2

